backend/requirements.txt
django==4.0.1
djangorestframework==3.13.1
dj-database-url==0.5.0
psycopg2==2.9.3
markdown==3.3.6
django-filter==21.1
django-environ==0.8.1